requirements.txt
# FastAPI & REST API
fastapi==0.111.1
uvicorn==0.23.2

# PDF & Document Processing
PyPDF2==3.0.1
pdfplumber==0.11.6
python-docx==1.2.0
python-multipart==0.0.20

# LLM & Vector DB
langchain==0.1.0
sentence-transformers==2.2.2
faiss-cpu==1.12.0
chromadb==1.1.1

# OpenAI / HuggingFace
openai==1.7.0
huggingface-hub==0.17.4
transformers==4.33.2

# Frontend (optional)
streamlit==1.25.0

# Environment variables
python-dotenv==1.0.1
pydantic-settings==2.11.0

# Database
SQLAlchemy==2.0.22
sqlmodel==0.0.8
databases==0.7.3
pydantic==2.2.1

# Utilities
requests==2.31.0
tqdm==4.66.1

# Testing
pytest==8.2.1
pytest-asyncio==0.22.0
httpx==0.25.1
